+++ log/docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- `if` statements now create a scope in `log.html` (when the function is not a generator).
 - `continue` and `break` inside a loop are properly handled.
 - `continue` and `break` are now shown in the logs.
 - Spec version changed from `0.0.3` to `0.0.4`.